refactor: restructure components and remoce unused stuff (clean npm run build)
