+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import Intro from "./intro";
+import Intro from "../components/landing/intro";
 import { useState, useEffect } from "react";
-import IntroPC from "./intropc";
+import IntroPC from "../components/landing/intropc";
 import Loading from "./loading";
 function HomePage() {
     const [isSlim, setSlim] = useState(false);
