+++ src/components/landing/button.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import { lato } from "./fonts";
 import React, { useEffect } from "react";
 import AOS from "aos";
 import "aos/dist/aos.css";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { stagger, useAnimate } from "framer-motion";
+import { lato } from "@/app/fonts";
 
 export default function Button() {
     const router = useRouter();
