+++ src/components/menu.tsx
@@ -2,9 +2,9 @@
 import React, { useState, useRef, useEffect } from "react";
 import AOS from "aos";
 import "aos/dist/aos.css";
-import "../../styles/menu.css";
+import "../styles/menu.css";
 import Link from "next/link";
-import { spectral } from "../fonts";
+import { spectral } from "@/app/fonts";
 
 type MenuProps = {
     link1: string;
